x86_64+kernel+common: refactor x86_64 to arch-consumes-kernel model
I sincerely apologize for the "everything, everywhere, all at once"
commit. This was the product of a design flaw realization I had
a few days ago that prompted me to shift the design of the kernel
around such that the architectures consumed the kernel instead of
the other way around, and to refactor things as quickly as possible.

It also contains a number of improvements:

- It significantly reduces the amount of boilerplace in commons.
- It significantly simplifies the bootloader process and further
  decouples the kernel and bootloader.
- Better separates out debugging facilities.
- Reduces memory footprint for the physical memory linear map.
- Reduces boot times on x86_64 by at least 10-20x.
- Slightly simplifies the linker story across architectures.
- Gives back control of many memory and performance aspects
  to the architecture instead of forcing them into a kernel-defined
  box.
- Isolates Oro-specific kernel functionality into a "safe" crate
  that remains architecture-agnostic.

This is a significant improvement to the overall structure of the
kernel and will undoubtedly have a huge impact on performance
and malleability of the codebase moving forward.

This commit has a few todo's attached to it that will be addressed
in further commits:

- This refactors a lot of the core/common crates as well as the kernel,
  but this commit only ports x86_64. Aarch64 will come in the next
  commit(s).
- No clippy checks have been addressed.
- Documentation is missing in a lot of places.
- The new boot process has not been documented or specified whatsoever.
- There is almost definitely not a lot of polish on the parts that
  were changed.

At the time of writing this is being developed on a feature branch
and will ultimately be merged to master now that the new design
is proven to be a huge improvement.